+++ src/index.js
@@ -6,7 +6,7 @@ export {default as cross} from "./cross.js";
 export {default as descending} from "./descending.js";
 export {default as deviation} from "./deviation.js";
 export {default as extent} from "./extent.js";
-export {default as group} from "./group.js";
+export {default as group, rollup} from "./group.js";
 export {default as bin, default as histogram} from "./bin.js"; // Deprecated; use bin.
 export {default as thresholdFreedmanDiaconis} from "./threshold/freedmanDiaconis.js";
 export {default as thresholdScott} from "./threshold/scott.js";
@@ -23,7 +23,6 @@ export {default as permute} from "./permute.js";
 export {default as quantile} from "./quantile.js";
 export {default as quickselect} from "./quickselect.js";
 export {default as range} from "./range.js";
-export {default as rollup} from "./rollup.js";
 export {default as least} from "./least.js";
 export {default as leastIndex} from "./leastIndex.js";
 export {default as scan} from "./scan.js"; // Deprecated; use leastIndex.
